Parallel deployments now are successfull
Enabled the capability for doing large deployments in a parallel fashion, this dramatically speeds doing multiple tasks across multiple hosts, and can be enabled by setting "parallel" : "true" for a deployment.

Also, fixed the action labels for the etcd action making it understandable what is taking place.

Finally, refactored the code around SSH to make it clearer where code should live.
